Pass cluster-name and refactor
