+++ .pre-commit-config.yaml
@@ -8,7 +8,7 @@ repos:
       - id: check-symlinks
 
   - repo: https://github.com/bridgecrewio/checkov.git
-    rev: 3.2.128
+    rev: 3.2.156
     hooks:
       - id: checkov
         files: Dockerfile
