+++ service.datadog.yaml
@@ -1,0 +1,21 @@
+schema-version: v2.2
+dd-service: gke-info
+team: platform-team
+description: >
+  Check the status of a GKE cluster. Currently, it only exposes /cluster-name and its primary goal is to test geolocation on multi-cluster deployments and ensure the Istio service mesh and location-based routing are working as expected.
+tier: Zero
+application: GKE Info
+languages:
+  - go
+links:
+  - name: Sandbox Global Platform Engineering
+    type: other
+    url: "https://sb.gcp.osinfra.io/gke-info"
+
+  - name: Pre-Production Global Platform Engineering
+    type: other
+    url: "https://pre-prod.gcp.osinfra.io/gke-info"
+
+  - name: Production Global Platform Engineering
+    type: other
+    url: "https://gcp.osinfra.io/gke-info"
